Use int for packet ids

+++ common/src/main/java/com/nukkitx/network/PacketCodec.java
@@ -8,5 +8,5 @@ public interface PacketCodec<T extends NetworkPacket> {
 
     ByteBuf tryEncode(T packet);
 
-    byte getId(T packet);
+    int getId(T packet);
 }
